+++ snake.js
@@ -1,7 +1,11 @@
 const cvs = document.getElementById("snake");
 const ctx = cvs.getContext("2d");
+const boardCVS = document.getElementById("board");
+const boardCTX = boardCVS.getContext("2d");
 let scoreElement = document.getElementById("score-number");
 let upgradeElement = document.getElementById("score-upgrade");
+let powerElement = document.getElementById("power-up");
+let time = document.getElementById("timer");
 
 //How many columns. There will be 0.8 times as many rows
 const COLUMNS = 25;
@@ -25,18 +29,29 @@ snake[0] = {
     y: (~~(COLUMNS / 2) - 2) * box
 }
 
+// Initializes power-up variables
+let halfSpeedX, halfSpeedY;
+halfSpeedX = halfSpeedY = -box;
+let smallX, smallY;
+smallX = smallY = -box;
+let pointsX, pointsY;
+pointsX = pointsY = -box;
+let extraX, extraY;
+extraX = extraY = -box;
+let extraOneX, extraOneY, extraTwoX, extraTwoY, extraThreeX, extraThreeY;
+extraOneX = extraOneY = extraTwoX = extraTwoY = extraThreeX = extraThreeY = -box;
+
 // Initializes taco location
 let foodX, foodY = 0;
 let foodIntersection = true;
 let foodIntersectCount = 0;
 while (foodIntersection == true) {
-    foodX = Math.floor(Math.random() * COLUMNS) * box;
-    foodY = Math.floor(Math.random() * (COLUMNS * 0.8)) * box;
-    for (let i = 0; i < snake.length; i++) {
-        if (snake[i].x == foodX && snake[i].y == foodY) {
-                foodIntersectCount++;
-            }
-    }
+    foodIntersectCount = 0;
+    foodX = ~~(Math.random() * COLUMNS) * box;
+    foodY = ~~(Math.random() * (COLUMNS * 0.8)) * box;
+    if (snake[0].x == foodX && snake[0].y == foodY) {
+            foodIntersectCount++;
+        }
     if (foodIntersectCount == 0) {
         foodIntersection = false;
     }
@@ -49,8 +64,50 @@ let food = {
 // Initializes half-speed power-up variables
 let halfSpeedActive = false;
 let halfSpeed = {
-    x: 0,
-    y: 0
+    x: -box,
+    y: -box
+}
+
+// Initializes -4 length power-up variables
+let smallActive = false;
+let small = {
+    x: -box,
+    y: -box
+}
+
+// Initializes bonus points power-up variables
+let pointsActive = false;
+let points = {
+    x: -box,
+    y: -box
+}
+
+// Initializes extra tacos power-up variables
+let extraActive = false;
+let extra = {
+    x: -box,
+    y: -box
+}
+
+// Initializes extra taco 1 variables
+let extraOneActive = false;
+let extraOne = {
+    x: -box,
+    y: -box
+}
+
+// Initializes extra taco 2 variables
+let extraTwoActive = false;
+let extraTwo = {
+    x: -box,
+    y: -box
+}
+
+// Initializes extra taco 3 variables
+let extraThreeActive = false;
+let extraThree = {
+    x: -box,
+    y: -box
 }
 
 // Initializes the score and score until speed increase variables
@@ -60,39 +117,56 @@ let scoreNeeded = UPGRADE;
 // Initializes variable to keep track of direction
 let d;
 
-document.addEventListener("keydown", direction);
+window.addEventListener("keydown", function(e) {
+    this.keyDown.call(this, e); 
+});
 
 // Determines which direction the player presses
-function direction(event) {
-    if (event.keyCode == 37 && d != "RIGHT") {
-        d = "LEFT"
+var keyBuffer = [];
+var keyDown = function(e) {
+    var keyCode = e.which ? e.which : e.keyCode;
+    // *** Queue the arrow key presses
+    if (keyCode >= 37 && keyCode <= 40 && keyCode !==
+        keyBuffer[keyBuffer.length - 1]) {
+        keyBuffer = keyBuffer.slice(-3).concat(keyCode);
     }
-    else if (event.keyCode == 38 && d != "DOWN") {
+};
+function direction() {
+    var key = keyBuffer.shift();
+    if (key == 37 && d != "RIGHT") {
+    d = "LEFT"
+    }
+    else if (key == 38 && d != "DOWN") {
         d = "UP"
     }
-    else if (event.keyCode == 39 && d != "LEFT") {
+    else if (key == 39 && d != "LEFT") {
         d = "RIGHT"
     }
-    else if (event.keyCode == 40 && d != "UP") {
+    else if (key == 40 && d != "UP") {
         d = "DOWN"
+    }
+}
+
+// Draws initial board
+function drawBoard() {
+    for (let i = 0; i < COLUMNS; i++){
+        for (let j = 0; j < ~~(COLUMNS * 0.8); j++){
+            if (i % 2 == j % 2) {
+                boardCTX.fillStyle = "purple";
+                boardCTX.fillRect(box * i , box * j, box, box);
+            }
+            else {
+                boardCTX.fillStyle = "pink";
+                boardCTX.fillRect(box * i, box * j, box, box);
+            }
+        }
     }
 }
 
 // Function for drawing all parts of the game and UI
 function draw() {
-    // Draws the board
-    for (let i = 0; i < COLUMNS; i++){
-        for (let j = 0; j < (COLUMNS * 0.8); j++){
-            if (i % 2 == j % 2) {
-                ctx.fillStyle = "purple";
-                ctx.fillRect(box * i , box * j, box, box);
-            }
-            else {
-                ctx.fillStyle = "pink";
-                ctx.fillRect(box * i, box * j, box, box);
-            }
-        }
-    }
+    // Clears the board
+    ctx.clearRect(0, 0, cvs.width, cvs.height);
 
     // Draws the snake
     for (let i = 0; i < snake.length; i++){
@@ -113,27 +187,34 @@ function draw() {
         score++;
         /* If score is a multiple of speed upgrade interval, the time is sped
         up by 20ms and the interval of the game is reset */
-        if (score % UPGRADE == 0 && SPEED >= 50) {
+        if (score % UPGRADE == 0 && SPEED >= 50 && scoreNeeded < 10) {
             SPEED -= 20;
             clearInterval(game);
             game = setInterval(draw, SPEED);
             scoreNeeded = UPGRADE;
         }
         else {
-            scoreNeeded = UPGRADE - (score % UPGRADE);
+            scoreNeeded--;
         }
 
         // Determines new location of taco
-        foodX, foodY = 0;
+        foodX, foodY = -box;
         foodIntersection = true;
         foodIntersectCount = 0;
         while (foodIntersection == true) {
-            foodX = Math.floor(Math.random() * COLUMNS) * box;
-            foodY = Math.floor(Math.random() * (COLUMNS * 0.8)) * box;
+            foodIntersectCount = 0;
+            foodX = ~~(Math.random() * COLUMNS) * box;
+            foodY = ~~(Math.random() * (COLUMNS * 0.8)) * box;
             for (let i = 0; i < snake.length; i++) {
                 if (snake[i].x == foodX && snake[i].y == foodY) {
-                        foodIntersectCount++;
-                    }
+                    foodIntersectCount++;
+                }
+            }
+            if ((foodX == extraOneX && foodY == extraOneY) || (foodX == extraTwoX && foodY == extraTwoY) ||
+                (foodX == extraThreeX && foodY == extraThreeY) || (foodX == halfSpeed.x && foodY == halfSpeed.y) ||
+                (foodX == small.x && foodY == small.y) || (foodX == points.x && foodY == points.y) ||
+                (foodX == extra.x && foodY == extra.y)) {
+                foodIntersectCount++;
             }
             if (foodIntersectCount == 0) {
                 foodIntersection = false;
@@ -149,20 +230,26 @@ function draw() {
         switch (powerChoice <= POWER_CHANCE) {
             // Case for half-speed power-up
             case (powerChoice <= (POWER_CHANCE / 4) && halfSpeedActive == false):
-                let halfSpeedX, halfSpeedY = 0;
-                let intersection = true;
-                let intersectCount = 0;
-                while (intersection == true) {
-                    halfSpeedX = Math.floor(Math.random() * COLUMNS) * box;
-                    halfSpeedY = Math.floor(Math.random() * (COLUMNS * 0.8)) * box;
+                halfSpeedX = halfSpeedY = -box;
+                let hsIntersection = true;
+                let hsIntersectCount = 0;
+                while (hsIntersection == true) {
+                    hsIntersectCount = 0;
+                    halfSpeedX = ~~(Math.random() * COLUMNS) * box;
+                    halfSpeedY = ~~(Math.random() * (COLUMNS * 0.8)) * box;
                     for (let i = 0; i < snake.length; i++){
-                        if ((snake[i].x == halfSpeedX && snake[i].y == halfSpeedY) ||
-                            (food.x == halfSpeedX && food.y == halfSpeedY)) {
-                            intersectCount++;
+                        if (snake[i].x == halfSpeedX && snake[i].y == halfSpeedY) {
+                            hsIntersectCount++;
                         }
                     }
-                    if (intersectCount == 0) {
-                        intersection = false;
+                    if ((halfSpeedX == extraOneX && halfSpeedY == extraOneY) || (halfSpeedX == extraTwoX && halfSpeedY == extraTwoY) ||
+                        (halfSpeedX == extraThreeX && halfSpeedY == extraThreeY) || (halfSpeedX == small.x && halfSpeedY == small.y) ||
+                        (halfSpeedX == points.x && halfSpeedY == points.y) || (halfSpeedX == food.x && halfSpeedY == food.y) ||
+                        (halfSpeedX == extra.x && halfSpeedY == extra.y)) {
+                        hsIntersectCount++;
+                    }
+                    if (hsIntersectCount == 0) {
+                        hsIntersection = false;
                     }
                 }
                 halfSpeed = {
@@ -171,12 +258,101 @@ function draw() {
                 }
                 halfSpeedActive = true;
                 break;
-            case (powerChoice <= ((POWER_CHANCE / 4) * 2)):
+            
+            // Case for -4 length power-up
+            case (powerChoice <= ((POWER_CHANCE / 4) * 2) && smallActive == false):
+                smallX = smallY = -box;
+                let smallIntersection = true;
+                let smallIntersectCount = 0;
+                while (smallIntersection == true) {
+                    smallIntersectCount = 0;
+                    smallX = ~~(Math.random() * COLUMNS) * box;
+                    smallY = ~~(Math.random() * (COLUMNS * 0.8)) * box;
+                    for (let i = 0; i < snake.length; i++){
+                        if (snake[i].x == smallX && snake[i].y == smallY) {
+                            smallIntersectCount++;
+                        }
+                    }
+                    if ((smallX == extraOneX && smallY == extraOneY) || (smallX == extraTwoX && smallY == extraTwoY) ||
+                        (smallX == extraThreeX && smallY == extraThreeY) || (smallX == halfSpeed.x && smallY == halfSpeed.y) ||
+                        (smallX == points.x && smallY == points.y) || (smallX == food.x && smallY == food.y) || 
+                        (smallX == extra.x && smallY == extra.y)) {
+                        smallIntersectCount++;
+                    }
+                    if (smallIntersectCount == 0) {
+                        smallIntersection = false;
+                    }
+                }
+                small = {
+                    x: smallX,
+                    y: smallY
+                }
+                smallActive = true;
                 break;
-            case (powerChoice <= ((POWER_CHANCE / 4) * 3)):
+            
+            // Case for bonus points power-up
+            case (powerChoice <= ((POWER_CHANCE / 4) * 3) && pointsActive == false):
+                pointsX = pointsY = -box;
+                let pointsIntersection = true;
+                let pointsIntersectCount = 0;
+                while (pointsIntersection == true) {
+                    pointsIntersectCount = 0;
+                    pointsX = ~~(Math.random() * COLUMNS) * box;
+                    pointsY = ~~(Math.random() * (COLUMNS * 0.8)) * box;
+                    for (let i = 0; i < snake.length; i++){
+                        if ((snake[i].x == pointsX && snake[i].y == pointsY) ||
+                            (food.x == pointsX && food.y == pointsY)) {
+                            pointsIntersectCount++;
+                        }
+                    }
+                    if ((pointsX == extraOneX && pointsY == extraOneY) || (pointsX == extraTwoX && pointsY == extraTwoY) ||
+                        (pointsX == extraThreeX && pointsY == extraThreeY) || (pointsX == halfSpeed.x && pointsY == halfSpeed.y) ||
+                        (pointsX == small.x && pointsY == small.y) || (pointsX == food.x && pointsY == food.y) || 
+                        (pointsX == extra.x && pointsY == extra.y)) {
+                        pointsIntersectCount++;
+                    }
+                    if (pointsIntersectCount == 0) {
+                        pointsIntersection = false;
+                    }
+                }
+                points = {
+                    x: pointsX,
+                    y: pointsY
+                }
+                pointsActive = true;
                 break;
+            
+            // Case for extra food power-up
             default:
-                
+                if (extraActive == false) {
+                    extraX = extraY = -box;
+                    let extraIntersection = true;
+                    let extraIntersectCount = 0;
+                    while (extraIntersection == true) {
+                        extraIntersectCount = 0;
+                        extraX = ~~(Math.random() * COLUMNS) * box;
+                        extraY = ~~(Math.random() * (COLUMNS * 0.8)) * box;
+                        for (let i = 0; i < snake.length; i++){
+                            if (snake[i].x == extraX && snake[i].y == extraY) {
+                                extraIntersectCount++;
+                            }
+                        }
+                        if ((extraX == extraOneX && extraY == extraOneY) || (extraX == extraTwoX && extraY == extraTwoY) ||
+                            (extraX == extraThreeX && extraY == extraThreeY) || (extraX == halfSpeed.x && extraY == halfSpeed.y) ||
+                            (extraX == small.x && extraY == small.y) || (extraX == food.x && extraY == food.y) ||
+                            (extraX == points.x && extraY == points.y)) {
+                            extraIntersectCount++;
+                        }
+                        if (extraIntersectCount == 0) {
+                            extraIntersection = false;
+                        }
+                    }
+                    extra = {
+                        x: extraX,
+                        y: extraY
+                    }
+                    extraActive = true;
+                }
         }
     }
     else {
@@ -190,15 +366,28 @@ function draw() {
     */
     if (snakeX == halfSpeed.x && snakeY == halfSpeed.y) {
         halfSpeedActive = false;
+        let counter = 15;
         let speedDifference = SPEED;
         SPEED *= 2;
         clearInterval(game);
         game = setInterval(draw, SPEED);
-        setTimeout(function () {
-            SPEED -= speedDifference;
-            clearInterval(game);
-            game = setInterval(draw, SPEED);
-        }, 15000);
+        let countDown = setInterval(function () {
+            time.classList.add('animation');
+            time.innerHTML = counter;
+            counter--;
+            if (counter < 0) {
+                time.innerHTML = "";
+                time.classList.remove('animation');
+                clearInterval(countDown);
+                SPEED -= speedDifference;
+                clearInterval(game);
+                game = setInterval(draw, SPEED);
+            }
+        }, 1000);
+        halfSpeed = {
+            x: -box,
+            y: -box
+        }    
     }
 
     // If half-speed power-up is in play, it will be drawn in a static position
@@ -207,11 +396,190 @@ function draw() {
         ctx.fillRect(halfSpeed.x, halfSpeed.y, box, box);
     }
 
-    // Updates score and "score until speedup" UI
+    /* Logic for colliding with -4 length power-up
+    Removes the last 4 tail pieces from the snake (if they exist)
+    */
+    if (snakeX == small.x && snakeY == small.y) {
+        smallActive = false;
+        for (let i = 1; i <= 4; i++) {
+            if (snake.length >= 1) {
+                snake.pop();
+            }
+        }
+        small = {
+            x: -box,
+            y: -box
+        }
+    }
+
+    // If -4 length power-up is in play, it will be drawn in a static position
+    if (smallActive == true) {
+        ctx.fillStyle = "blue";
+        ctx.fillRect(small.x, small.y, box, box);
+    }
+
+    /* Logic for colliding with bonus points power-up
+    Adds 10 points to current score without incrementing speed of game
+    */
+    if (snakeX == points.x && snakeY == points.y) {
+        pointsActive = false;
+        score += 10;
+        points = {
+            x: -box,
+            y: -box
+        }
+    }
+
+    // If bonus points power-up is in play, it will be drawn in a static position
+    if (pointsActive == true) {
+        ctx.fillStyle = "orange";
+        ctx.fillRect(points.x, points.y, box, box);
+    }
+
+    /* Logic for colliding with extra tacos power-up
+    Spawns three golden tacos
+    */
+    if (snakeX == extra.x && snakeY == extra.y) {
+        extraActive = false;
+        extraOneX = extraOneY = extraTwoX = extraTwoY = extraThreeX = extraThreeY = -box;
+        let extraOneIntersection, extraTwoIntersection, extraThreeIntersection;
+        extraOneIntersection = extraTwoIntersection = extraThreeIntersection = true;
+        let extraOneIntersectCount, extraTwoIntersectCount, extraThreeIntersectCount;
+        extraOneIntersectCount = extraTwoIntersectCount = extraThreeIntersectCount = 0;
+        while (extraOneIntersection == true) {
+            extraOneIntersectCount = 0;
+            extraOneX = ~~(Math.random() * COLUMNS) * box;
+            extraOneY = ~~(Math.random() * (COLUMNS * 0.8)) * box;
+            for (let i = 0; i < snake.length; i++){
+                if (snake[i].x == extraOneX && snake[i].y == extraOneY) {
+                    extraOneIntersectCount++;
+                }
+            }
+            if ((extraOneX == extra.x && extraOneY == extra.y) || (extraOneX == extraTwoX && extraOneY == extraTwoY) ||
+                (extraOneX == extraThreeX && extraOneY == extraThreeY) || (extraOneX == halfSpeed.x && extraOneY == halfSpeed.y) ||
+                (extraOneX == small.x && extraOneY == small.y) || (extraOneX == food.x && extraOneY == food.y) ||
+                (extraOneX == points.x && extraOneY == points.y)) {
+                extraOneIntersectCount++;
+            }
+            if (extraOneIntersectCount == 0) {
+                extraOneIntersection = false;
+            }
+        }
+        extraOne = {
+            x: extraOneX,
+            y: extraOneY
+        }
+        
+        extraOneActive = true;
+        while (extraTwoIntersection == true) {
+            extraTwoIntersectCount = 0;
+            extraTwoX = ~~(Math.random() * COLUMNS) * box;
+            extraTwoY = ~~(Math.random() * (COLUMNS * 0.8)) * box;
+            for (let i = 0; i < snake.length; i++){
+                if (snake[i].x == extraTwoX && snake[i].y == extraTwoY) {
+                    extraTwoIntersectCount++;
+                }
+            }
+            if ((extraTwoX == extraOneX && extraTwoY == extraOneY) || (extraTwoX == extra.x && extraTwoY == extra.y) ||
+                (extraTwoX == extraThreeX && extraTwoY == extraThreeY) || (extraTwoX == halfSpeed.x && extraTwoY == halfSpeed.y) ||
+                (extraTwoX == small.x && extraTwoY == small.y) || (extraTwoX == food.x && extraTwoY == food.y) ||
+                (extraTwoX == points.x && extraTwoY == points.y)) {
+                extraTwoIntersectCount++;
+            }
+            if (extraTwoIntersectCount == 0) {
+                extraTwoIntersection = false;
+            }
+        }
+        extraTwo = {
+            x: extraTwoX,
+            y: extraTwoY
+        }
+        extraTwoActive = true;
+        while (extraThreeIntersection == true) {
+            extraThreeIntersectCount = 0;
+            extraThreeX = ~~(Math.random() * COLUMNS) * box;
+            extraThreeY = ~~(Math.random() * (COLUMNS * 0.8)) * box;
+            for (let i = 0; i < snake.length; i++){
+                if (snake[i].x == extraThreeX && snake[i].y == extraThreeY) {
+                    extraThreeIntersectCount++;
+                }
+            }
+            if ((extraThreeX == extraOneX && extraThreeY == extraOneY) || (extraThreeX == extraTwoX && extraThreeY == extraTwoY) ||
+                (extraThreeX == extra.x && extraThreeY == extra.y) || (extraThreeX == halfSpeed.x && extraThreeY == halfSpeed.y) ||
+                (extraThreeX == small.x && extraThreeY == small.y) || (extraThreeX == food.x && extraThreeY == food.y) ||
+                (extraThreeX == points.x && extraThreeY == points.y)) {
+                extraThreeIntersectCount++;
+            }
+            if (extraThreeIntersectCount == 0) {
+                extraThreeIntersection = false;
+            }
+        }
+        extraThree = {
+            x: extraThreeX,
+            y: extraThreeY
+        }
+        extraThreeActive = true;
+    }
+
+    // If extra tacos power-up is in play, it will be drawn in a static position
+    if (extraActive == true) {
+        ctx.fillStyle = "gray";
+        ctx.fillRect(extra.x, extra.y, box, box);
+    }
+
+    /* Logic for colliding with golden tacos
+    Increase score by 2 and increase tacos until speed-up by 1 per taco
+    */
+    if (snakeX == extraOne.x && snakeY == extraOne.y || snakeX == extraTwo.x && snakeY == extraTwo.y ||
+        snakeX == extraThree.x && snakeY == extraThree.y) {
+        score += 2;
+        scoreNeeded += 1;
+        if (snakeX == extraOne.x && snakeY == extraOne.y) {
+            extraOneActive = false;
+            upgradeElement.innerHTML = scoreNeeded + " tacos until speedup";
+            extraOne = {
+                x: -box,
+                y: -box
+            }
+        }
+        else if (snakeX == extraTwo.x && snakeY == extraTwo.y) {
+            extraTwoActive = false;
+            upgradeElement.innerHTML = scoreNeeded + " tacos until speedup";
+            extraTwo = {
+                x: -box,
+                y: -box
+            }
+        }
+        else {
+            extraThreeActive = false;
+            upgradeElement.innerHTML = scoreNeeded + " tacos until speedup";
+            extraThree = {
+                x: -box,
+                y: -box
+            }
+        }
+    }
+
+    // If golden tacos are in play, it will be drawn in a static position
+    if (extraOneActive == true || extraTwoActive == true || extraThreeActive == true) {
+        ctx.fillStyle = "black";
+        if (extraOneActive == true) {
+            ctx.fillRect(extraOne.x, extraOne.y, box, box);
+        }
+        if (extraTwoActive == true) {
+            ctx.fillRect(extraTwo.x, extraTwo.y, box, box);
+        }
+        if (extraThreeActive == true) {
+            ctx.fillRect(extraThree.x, extraThree.y, box, box);
+        }
+    }
+
+    // Updates score, "score until speedup", and power-up UI
     scoreElement.innerHTML = score;
     upgradeElement.innerHTML = scoreNeeded + " tacos until speedup";
 
     // Determines how to move snake depending on the direction it is moving
+    direction();
     if (d == "LEFT") {
         snakeX -= box;
     }
@@ -251,4 +619,5 @@ function draw() {
 }
 
 // Starts game
+drawBoard();
 let game = setInterval(draw, SPEED);